% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggressor.relation.code.R
\docType{data}
\name{aggressor.relation.code}
\alias{aggressor.relation.code}
\title{Aggressor Relationship codes}
\format{A data frame with 74 observations on the following 2 variables.}
\usage{
aggressor.relation.code
}
\description{
This dataset contains all codes defining the relationship (if any) between the aggressor 
and their victim for the aggressor_relation_code column in the injury intent database
\url{http://www.inegi.org.mx/est/contenidos/proyectos/registros/vitales/mortalidad/default.aspx}
}
\section{Variables}{


\itemize{
\item{\code{aggressor_relation_code}}{a numeric vector}
\item{\code{relationship}}{a character vector}
}
}

\examples{
head(aggressor.relation.code)
}
