% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.codes.R
\docType{data}
\name{geo.codes}
\alias{geo.codes}
\title{State and Municipio Codes}
\format{A data frame with 2559 observations on the following 3 variables.}
\usage{
geo.codes
}
\description{
This dataset contains all state and municipio (county) codes corresponding to those
in the injury intent mortality database
\url{http://www.inegi.org.mx/est/contenidos/proyectos/registros/vitales/mortalidad/default.aspx}
}
\section{Variables}{


\itemize{
\item{\code{state_code}}{a numeric vector}
\item{\code{mun_code}}{a numeric vector}
\item{\code{name}}{a character vector}
}
}
\examples{
head(geo.codes)
}

