\docType{data}
\name{icd.103}
\alias{icd.103}
\title{List of 103 Diseases}
\format{A data frame with 103 observations on the following 2 variables.}
\usage{
icd.103
}
\description{
This dataset contains a list of 103 diseases to match with
the injury mortality dababse variable icd_103
\url{http://www.inegi.org.mx/est/contenidos/proyectos/registros/vitales/mortalidad/default.aspx}
}
\section{Variables}{
  \itemize{ \item{\code{icd_103}}{: a numeric vector}
  \item{\code{icd_103_name}}{: a character vector} }
}
\examples{
head(icd.103)
}

