\docType{data}
\name{mex.list.group}
\alias{mex.list.group}
\title{Dicease Group corresponding to the Mexican Dicease List}
\format{A data frame with 59 observations on the following 2 variables.}
\description{
  This dataset contains a list of disease groups to match
  with the injury mortality database variable
  mex_list_group
  \url{http://www.inegi.org.mx/est/contenidos/proyectos/registros/vitales/mortalidad/default.aspx}
}
\section{Variables}{
  \itemize{ \item{\code{mex_list_group}}{: a character
  vector} \item{\code{mex_list_group_name}}{: a character
  vector} }
}
\examples{
head(mex.list.group)
}
\keyword{datasets}

