% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mex.list.R
\docType{data}
\name{mex.list}
\alias{mex.list}
\title{Mexican Disease List}
\format{A data frame with 417 observations on the following 2 variables.}
\usage{
mex.list
}
\description{
This dataset contains a list of diseases to match with the injury mortality database variable mex_list
\url{http://www.inegi.org.mx/est/contenidos/proyectos/registros/vitales/mortalidad/default.aspx}
}
\section{Variables}{


\itemize{
\item{\code{mex_list}}{: a character vector}
\item{\code{mex_list_name}}{: a character vector}
}
}

\examples{
head(mex.list)
}
